+++ tests/test_app.py
@@ -5,6 +5,7 @@ import responses
 
 import azure_costs_exporter
 from azure_costs_exporter.main import create_app
+from azure_costs_exporter.views import DEFAULT_SCRAPE_TIMEOUT
 from .data import sample_data, api_output_for_empty_months
 
 
@@ -44,23 +45,35 @@ def test_token(client, now, enrollment, access_key):
         json=sample_data
     )
 
-    assert client.get('/metrics').status_code == 200
+    rv = client.get('/metrics')
+    assert rv.status_code == 200
     assert responses.calls[-1].request.headers['Authorization'] == "Bearer {}".format(access_key)
 
 
-@responses.activate
-def test_metrics(app, now, enrollment):
-    
-    responses.add(
-        method='GET',
-        url="https://ea.azure.com/rest/{0}/usage-report?month={1}&type=detail&fmt=Json".format(enrollment, now),
-        match_querystring=True,
-        json=sample_data
-    )
+@pytest.mark.parametrize('timeout,expected', [('42.3', 42.3), (None, DEFAULT_SCRAPE_TIMEOUT)])
+def test_metrics(app, access_key, now, enrollment, timeout, expected):
+    class RequestsMock(responses.RequestsMock):
+        def get(self, *args, **kwargs):
+            assert kwargs['timeout'] == expected
+            return super(RequestsMock, self).get(*args, **kwargs)
 
-    rsp = app.test_client().get('/metrics')
-    assert rsp.status_code == 200
-    assert rsp.data.count(b'azure_costs_eur') == 4
+    with RequestsMock() as resp:
+        resp.add(
+            method='GET',
+            url="https://ea.azure.com/rest/{0}/usage-report?month={1}&type=detail&fmt=Json".format(enrollment, now),
+            match_querystring=True,
+            json=sample_data
+        )
+
+        headers = {}
+        if timeout is not None:
+            headers = {'X-Prometheus-Scrape-Timeout-Seconds': timeout}
+
+        rsp = app.test_client().get('/metrics', headers=headers)
+        url = 'https://ea.azure.com/rest/{enrollment}/usage-report?month={month}&type=detail&fmt=Json'
+        url = url.format(enrollment=enrollment, month=now)
+        assert rsp.status_code == 200
+        assert rsp.data.count(b'azure_costs_eur') == 4
 
 
 @responses.activate
@@ -79,25 +92,6 @@ def test_metrics_no_usage(app, now, enrollment):
 
     # expect only metric definition and help but no content in the output
     assert rsp.data.count(b'azure_costs_eur') == 2
-
-
-@responses.activate
-def test_metrics_timeout(monkeypatch, app, access_key, now, enrollment):
-    get_mock = MagicMock()
-    monkeypatch.setattr(azure_costs_exporter.prometheus_collector.requests,
-                        'get',
-                        get_mock)
-
-    rsp = app.test_client().get('/metrics', headers={'Scrape-Timeout-Seconds': 42.0})
-    url = 'https://ea.azure.com/rest/{enrollment}/usage-report?month={month}&type=detail&fmt=Json'
-    url = url.format(enrollment=enrollment, month=now)
-
-    get_mock.assert_called_once_with(
-        url,
-        headers={'Authorization': 'Bearer {}'.format(access_key)},
-        timeout=42
-    )
-    assert rsp.status_code == 200
 
 
 @responses.activate